#7446: skip the voids the rest of a copy chain already filled
