+++ eo-inference/src/main/java/org/eolang/inference/Bound.java
@@ -4,6 +4,10 @@
  */
 package org.eolang.inference;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -16,7 +20,10 @@ import java.util.Map;
  * {@code t}. The place of an argument says which void it lands in, counted
  * through the voids in the order the formation declares them, which is the
  * order {@link Ungrouped} keeps, past the voids the rest of the chain of
- * copies already took.</p>
+ * copies already took: {@code half 2 > full} copies {@code half}, which is a
+ * copy of {@code pair}, and the void {@code half} filled is gone by the time
+ * {@code full} applies, the way {@code PhDefault.vacancy()} walks past an
+ * attribute that already holds something.</p>
  *
  * <p>What fills the void is written down as the argument itself and not as
  * what the argument turns out to be. The two differ exactly when the argument
@@ -35,9 +42,19 @@ final class Bound {
     private final Map<String, List<String>> args;
 
     /**
-     * What every application in a copy chain has already filled.
+     * Every name against the one it is a direct copy of.
      */
-    private final Consumed consumed;
+    private final Map<String, String> chain;
+
+    /**
+     * The name every type goes by.
+     */
+    private final Map<String, String> names;
+
+    /**
+     * What the types certainly have.
+     */
+    private final Provided owned;
 
     /**
      * Ctor.
@@ -51,7 +68,9 @@ final class Bound {
         final Provided provided
     ) {
         this.args = arguments;
-        this.consumed = new Consumed(arguments, copies, provided);
+        this.chain = copies;
+        this.names = new Ends(copies).names();
+        this.owned = provided;
     }
 
     /**
@@ -63,11 +82,46 @@ final class Bound {
     Map<String, Map<String, String>> all() {
         final Map<String, Map<String, String>> found = new LinkedHashMap<>(0);
         for (final String application : this.args.keySet()) {
-            final Map<String, String> filled = this.consumed.filled(application);
+            final Map<String, String> filled = this.filled(application);
             if (!filled.isEmpty()) {
                 found.put(application, filled);
             }
         }
         return found;
+    }
+
+    private Map<String, String> filled(final String application) {
+        final Collection<String> taken = new HashSet<>(0);
+        Map<String, String> mine = new LinkedHashMap<>(0);
+        for (final String step : this.order(application)) {
+            mine = this.stepped(step, taken);
+            taken.addAll(mine.keySet());
+        }
+        return mine;
+    }
+
+    private Map<String, String> stepped(final String step, final Collection<String> taken) {
+        final Map<String, String> mine = new LinkedHashMap<>(0);
+        final String copied = this.names.getOrDefault(step, step);
+        final List<String> given = this.args.get(step);
+        for (int place = 0; place < given.size(); place += 1) {
+            final String hollow = this.owned.slot(copied, place, taken);
+            if (!hollow.isEmpty() && !given.get(place).isEmpty()) {
+                mine.put(hollow, given.get(place));
+            }
+        }
+        return mine;
+    }
+
+    private List<String> order(final String node) {
+        final List<String> steps = new ArrayList<>(0);
+        final Collection<String> seen = new HashSet<>(0);
+        String walked = node;
+        while (this.args.containsKey(walked) && seen.add(walked)) {
+            steps.add(walked);
+            walked = this.chain.getOrDefault(walked, "");
+        }
+        Collections.reverse(steps);
+        return steps;
     }
 }
